Revert "Improve demo mode"
