+++ browser_use/agent/service.py
@@ -930,6 +930,8 @@ class Agent(Generic[Context, AgentStructuredOutput]):
 
 		# Log step completion summary
 		summary_message = self._log_step_completion_summary(self.step_start_time, self.state.last_result)
+		if summary_message:
+			await self._demo_mode_log(summary_message, 'info', {'step': self.state.n_steps})
 
 		# Save file system state after step completion
 		self.save_file_system_state()
@@ -1767,6 +1769,12 @@ class Agent(Generic[Context, AgentStructuredOutput]):
 		"""
 		if on_step_start is not None:
 			await on_step_start(self)
+
+		await self._demo_mode_log(
+			f'Starting step {step + 1}/{max_steps}',
+			'info',
+			{'step': step + 1, 'total_steps': max_steps},
+		)
 
 		self.logger.debug(f'🚶 Starting step {step + 1}/{max_steps}...')
